# api/modules/linux/openapi.yaml
openapi: 3.1.0
...
info:
  title: Opale Linux fleet API
  version: 1.0.0
  description: |
    Control plane for Debian workstations managed by pull: devices enrol with a self-generated Ed25519
    key, are approved (or auto-approved by serial pre-registration), fetch their assignment at each
    check-in, pull their goal state from Opale's read-only git mirror with `ansible-pull`, and report
    the result. Opale never runs code on a Linux device: it only chooses which *signed* revision each
    ring applies.

    ## Device authentication (all `/api/linux/agent/*` operations except the git smart-HTTP routes)
    Every request carries:
    - `x-opale-key`: device key fingerprint = hex(sha256(raw 32-byte Ed25519 public key))
    - `x-opale-timestamp`: unix seconds
    - `x-opale-nonce`: 16 random bytes, hex
    - `x-opale-signature`: base64(Ed25519.sign(private_key, canonical))

    `canonical = "opale-linux-v1\n" + METHOD + "\n" + path_with_query + "\n" + timestamp + "\n" + nonce + "\n" + hex(sha256(raw_request_body))`
    (`path_with_query` is the request target as sent, e.g. `/api/linux/agent/checkin`; an empty body hashes as sha256 of zero bytes.)

    Server checks, in order: key row exists (on `/enroll` any status; elsewhere status must be `approved`),
    `|now - timestamp| <= 300 s` (else 401 `CLOCK_SKEW` with `server_time`), nonce unseen in the last 10
    minutes, signature valid over the exact bytes received. Failures are 401 with a `code`.

    ## Git smart HTTP
...

...
servers:
...
security:
  - deviceSignature: []
paths:
...
  /linux/agent/enroll:
    post:
      operationId: linuxAgentEnroll
      tags: [linux-agent]
      summary: First contact / poll for approval (idempotent)
      description: |
        Signed with the device key even though the key is unknown to the server on first contact.
        The server derives the fingerprint from `public_key` and requires it to equal `x-opale-key`.

        Outcomes:
        - unknown key → a `pending` enrolment is created (`202`), unless the normalized serial matches a
          pre-registration AND no existing device / approved key carries that serial → auto-approved (`200`).
        - serial matches an existing device (any platform) or an approved key → `202` with `conflict`,
          the pre-registration is NOT consumed; an admin decides (convert / supersede / reject).
        - pending key → `202` (retry after `retry_after_s`); approved → `200`; rejected/revoked → `403`.
        - approved key presenting a different normalized serial → `403 SERIAL_MISMATCH`.
        - more than 500 pending enrolments → `429` (`linux_enroll_flood` audited); re-contacts of existing
          keys are never blocked.
      x-opale-rate-limit: "30/min per IP (ipOnlyKey); never keyed by fingerprint"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnrollRequest' }
      responses:
        '200':
          description: Approved — the device may call `/checkin`.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnrollApproved' } } }
        '202':
          description: Pending approval.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnrollPending' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Rejected, revoked, or serial mismatch.
          content: { application/json: { schema: { $ref: '#/components/schemas/EnrollRefused' } } }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /linux/agent/checkin:
    post:
      operationId: linuxAgentCheckin
      tags: [linux-agent]
      summary: Report liveness and fetch the assignment
      description: |
        Writes `devices.last_seen`, `devices.os`, `devices.disk_used_pct` and the key row's
        `last_seen_at/agent_version/os_version`. Never writes `devices.hostname` or `devices.agent_version`.
        `serial` is compared (normalized) with the enrolled serial: a mismatch is refused with
        `403 SERIAL_MISMATCH` and audited (`linux_key_serial_mismatch`, throttled). A NULL/placeholder
        enrolled serial adopts the first real serial seen.

        The agent applies `hostname`, then runs immediately (same cycle):
        `ansible-pull -U <git.url> -C <ring> --verify-commit --only-if-changed -e opale_profile=… profiles/<profile>.yml`
        with `GIT_TERMINAL_PROMPT=0` and `GIT_CONFIG_*` = `http.<git.url>.extraHeader`, `gpg.format=ssh`,
        `gpg.ssh.allowedSignersFile=/etc/opale-agent/allowed_signers`. If the git stage fails, wait 30 s,
        check in again (fresh token) and retry once before reporting `failed`. `git` is `null` while the
        mirror is not ready: skip the apply, do not report, retry after `retry_after_s`.
      x-opale-rate-limit: "IP 60/min; 12/hour per device enforced in the handler"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CheckinRequest' }
      responses:
        '200':
          description: The assignment.
          content: { application/json: { schema: { $ref: '#/components/schemas/Assignment' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Serial mismatch.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /linux/agent/reports:
    post:
      operationId: linuxAgentReport
      tags: [linux-agent]
      summary: Report the outcome of an ansible-pull run
      description: |
        Appends to `linux_apply_reports` (append-only, 90-day retention). `log_tail` is discarded when
        `status` is `success` and truncated (head+tail) to 8 KiB otherwise. `skipped` = `--only-if-changed`
        found nothing to do, or "no hosts matched"; send it without a log. Updates
        `devices.last_revision_applied / last_successful_revision / last_apply_status / last_apply_at`.
        Only transitions are audited (`linux_apply_failed`, `linux_apply_recovered`); a failure may also
        raise a push notification and a ticket proposal when `linux.alerts_enabled` is on.
      x-opale-rate-limit: "IP 60/min; 12/hour per device enforced in the handler"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReportRequest' }
      responses:
        '201':
          description: Stored.
          content: { application/json: { schema: { $ref: '#/components/schemas/ReportAck' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /linux/agent/escrow:
    post:
      operationId: linuxAgentEscrow
      tags: [linux-agent]
      summary: Escrow a secret encrypted to the instance escrow key
      description: |
        `ciphertext` is base64 of RSA-OAEP-SHA256(plaintext) under `escrow.public_key_pem` from the last
        check-in (decoded size 200..1024 bytes; keep plaintext ≤ 150 bytes). `escrow_key_id` must equal the
        served `escrow.key_id` (else `409 ESCROW_KEY_MISMATCH`).

        - `local_admin`: `username` must equal the assignment's `local_admin_username` (else
          `400 USERNAME_MISMATCH`); `rotated_at` must be later than the stored `password_changed_at`
          (else `409 ESCROW_STALE`) and at most 5 min in the future (else `400 ROTATED_AT_IN_FUTURE`).
          Upserts the device's row in `device_admin_credentials` (same table and reveal route as
          Windows LAPS), clears `rotation_requested_at`, audits `laps_rotated`.
        - `luks_recovery`: inserts a new `device_recovery_keys` row; a previous row with the same `label`
          is marked superseded (history kept). Refused with `409 ESCROW_BACKUP_UNCONFIRMED` until an admin
          has confirmed the off-box backup of the escrow key (`POST /linux/escrow/confirm-backup`).

        Escrow first, then apply locally (the agent must never hold a secret the server does not).
      x-opale-rate-limit: "IP 60/min; 6/min per device enforced in the handler"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EscrowRequest' }
      responses:
        '201':
          description: Stored.
          content: { application/json: { schema: { $ref: '#/components/schemas/EscrowAck' } } }
        '400':
          description: Validation error, CIPHERTEXT_SIZE (decoded size outside 200..1024), USERNAME_MISMATCH or ROTATED_AT_IN_FUTURE.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: ESCROW_KEY_MISMATCH, ESCROW_STALE or ESCROW_BACKUP_UNCONFIRMED.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '503':
          description: ESCROW_UNAVAILABLE — the escrow private key is unreadable on the server.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
components:
  securitySchemes:
    deviceSignature:
      type: apiKey
      in: header
      name: x-opale-signature
      description: Ed25519 signed request (see the top-level description); also requires x-opale-key, x-opale-timestamp, x-opale-nonce.
    gitToken:
      type: http
      scheme: bearer
      description: Short-lived git token from `/checkin` (`git.token`).
    entraJwt:
      type: http
      scheme: bearer
      description: Entra ID access token or `opl_` CLI token. Operations marked `x-opale-interactive` refuse CLI tokens (403 INTERACTIVE_ONLY).
  responses:
...
    Unauthorized:
      description: Missing/invalid credential. Device codes — SIGNATURE_INVALID, UNKNOWN_KEY, CLOCK_SKEW (with server_time), NONCE_REPLAY, REVOKED, NOT_APPROVED; git — TOKEN_INVALID, TOKEN_EXPIRED.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    TooManyRequests:
      description: Rate limited; `Retry-After` header and `retry_after_ms`.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: "French, user-facing" }
        code: { type: string, description: "Machine-readable, UPPER_SNAKE" }
        server_time: { type: integer, description: "unix seconds; present on CLOCK_SKEW" }
        retry_after_ms: { type: integer }
        details: { type: object, additionalProperties: true }
      examples:
        - { error: "Horloge du poste décalée", code: CLOCK_SKEW, server_time: 1790800000 }
        - { error: "Poste géré par état désiré", code: PULL_MANAGED }
...
    Ring:
      type: string
      enum: [pilot, stable]
    ProfileSlug:
      type: string
      pattern: '^[a-z0-9][a-z0-9-]{0,63}$'
      description: "`profiles/<slug>.yml` in the fleet repo"
      examples: [ field-researcher, admin ]
    Hostname:
      type: string
      pattern: '^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$'
      description: RFC-1123 label, lowercase
    Serial:
      type: [string, 'null']
      maxLength: 100
      description: SMBIOS product serial as read from DMI; normalized server-side (trim, uppercase; placeholders such as "To be filled by O.E.M." become null)
...
    ApplyStatus:
      type: string
      enum: [success, failed, partial, skipped]
    KeyBacking:
      type: string
      enum: [software, tpm]
...
    EnrollRequest:
      type: object
      additionalProperties: false
      required: [serial, hostname, os_version, agent_version, key_backing, public_key]
      properties:
        serial: { $ref: '#/components/schemas/Serial' }
        hostname: { type: string, maxLength: 253, description: "current hostname (claim, display only)" }
        os_version: { type: string, maxLength: 100, description: "PRETTY_NAME or VERSION_ID from /etc/os-release" }
        kernel: { type: string, maxLength: 100 }
        agent_version: { type: string, maxLength: 40 }
        key_backing: { $ref: '#/components/schemas/KeyBacking' }
        public_key: { type: string, pattern: '^[A-Za-z0-9+/]{43}=$', description: "base64 of the raw 32-byte Ed25519 public key" }
      examples:
        - serial: "PF3ABC12"
          hostname: debian
          os_version: "Debian GNU/Linux 12 (bookworm)"
          kernel: "6.1.0-25-amd64"
          agent_version: "0.1.0"
          key_backing: software
          public_key: "Gb9ECWmEzf6FQbrBZ9w7lshQhqowtrbLDFw4rXAxZuE="
    EnrollPending:
      type: object
      required: [status, code, retry_after_s]
      properties:
        status: { type: string, const: pending }
        code: { type: string, pattern: '^[0-9a-f]{8}$', description: "first 8 hex of the fingerprint — shown on the device screen so an admin can find it in the queue" }
        retry_after_s: { type: integer, minimum: 10, description: "30 s during the first hour after first contact, then 300 s" }
        conflict:
          type: [object, 'null']
          required: [kind]
          properties:
            kind: { type: string, enum: [existing_device, reimage, clone] }
      examples:
        - { status: pending, code: "3f9a1c22", retry_after_s: 30, conflict: null }
    EnrollApproved:
      type: object
      required: [status, device_id, hostname]
      properties:
        status: { type: string, const: approved }
        device_id: { type: string, format: uuid }
        hostname: { $ref: '#/components/schemas/Hostname' }
      examples:
        - { status: approved, device_id: "7f9c2a4e-3b1d-4c6e-9a8f-1d2e3f4a5b6c", hostname: lx-pf3abc12 }
    EnrollRefused:
      type: object
      required: [status, error]
      properties:
        status: { type: string, enum: [rejected, revoked, serial_mismatch] }
        error: { type: string }
        code: { type: string }
      examples:
        - { status: revoked, error: "Clé révoquée", code: REVOKED }
    CheckinRequest:
      type: object
      additionalProperties: false
      required: [serial, agent_version, os_version, hostname, luks_root]
      properties:
        serial: { $ref: '#/components/schemas/Serial' }
        agent_version: { type: string, maxLength: 40 }
        os_version: { type: string, maxLength: 100 }
        kernel: { type: string, maxLength: 100 }
        hostname: { type: string, maxLength: 253 }
        uptime_s: { type: integer, minimum: 0 }
        disk_root_pct: { type: number, minimum: 0, maximum: 100 }
        luks_root: { type: boolean, description: "root filesystem is on a LUKS volume" }
        escrow_key_id: { type: [string, 'null'], pattern: '^[0-9a-f]{64}$', description: "key_id of the last successful escrow, null if never" }
        local_admin_rotated_at: { type: [string, 'null'], format: date-time }
      examples:
        - serial: "PF3ABC12"
          agent_version: "0.1.0"
          os_version: "Debian GNU/Linux 12 (bookworm)"
          kernel: "6.1.0-25-amd64"
          hostname: lx-pf3abc12
          uptime_s: 86400
          disk_root_pct: 41.5
          luks_root: true
          escrow_key_id: null
          local_admin_rotated_at: null
    Assignment:
      type: object
      required: [device_id, hostname, profile, ring, revision, git, escrow, escrow_needed, local_admin_username, rotate_local_admin, checkin_interval_s, extra_vars]
      properties:
        device_id: { type: string, format: uuid }
        hostname: { $ref: '#/components/schemas/Hostname' }
        profile: { $ref: '#/components/schemas/ProfileSlug' }
        ring: { $ref: '#/components/schemas/Ring' }
        revision: { type: [string, 'null'], pattern: '^[0-9a-f]{40}$', description: "what the ring currently resolves to (informational; the agent checks out the ring)" }
        git:
          type: [object, 'null']
          required: [url, token, token_expires_at]
          properties:
            url: { type: string, format: uri, description: "must equal <configured server>/api/linux/agent/git/fleet.git" }
            token: { type: string }
            token_expires_at: { type: string, format: date-time }
        retry_after_s: { type: integer, description: "present when git is null (mirror not ready)" }
        escrow:
          type: object
          required: [status]
          properties:
            status: { type: string, enum: [ok, unavailable, backup_unconfirmed] }
            public_key_pem: { type: [string, 'null'] }
            key_id: { type: [string, 'null'], pattern: '^[0-9a-f]{64}$' }
        escrow_needed:
          type: array
          items: { type: string, enum: [local_admin, luks_recovery] }
          description: what the server is missing for this device under the current key_id
        local_admin_username: { type: string }
        rotate_local_admin: { type: boolean, description: "an admin requested a rotation" }
        checkin_interval_s: { type: integer, minimum: 60 }
        extra_vars:
          type: object
          required: [opale_profile, opale_ring, opale_device_id, opale_hostname]
          properties:
            opale_profile: { type: string }
            opale_ring: { type: string }
            opale_device_id: { type: string }
            opale_hostname: { type: string }
      examples:
        - device_id: "7f9c2a4e-3b1d-4c6e-9a8f-1d2e3f4a5b6c"
          hostname: lx-pf3abc12
          profile: field-researcher
          ring: stable
          revision: "b89f176e2c1c4d1f0a9e8d7c6b5a49382716f5e4"
          git:
            url: "https://opale.example.org/api/linux/agent/git/fleet.git"
            token: "gt_9c1f0a7e3b2d4c5f6a8b9c0d1e2f3a4b5c6d7e"
            token_expires_at: "2026-09-30T20:10:00Z"
          escrow: { status: ok, public_key_pem: "-----BEGIN PUBLIC KEY-----\nMIIB…\n-----END PUBLIC KEY-----\n", key_id: "0f3c4d5e6f7a8b9c0d1e2f3a4b5c6d7e8f9a0b1c2d3e4f5a6b7c8d9e0f1a2b3c" }
          escrow_needed: [luks_recovery]
          local_admin_username: opale-admin
          rotate_local_admin: false
          checkin_interval_s: 900
          extra_vars: { opale_profile: field-researcher, opale_ring: stable, opale_device_id: "7f9c2a4e-3b1d-4c6e-9a8f-1d2e3f4a5b6c", opale_hostname: lx-pf3abc12 }
    ReportRequest:
      type: object
      additionalProperties: false
      required: [revision, status, started_at, finished_at]
      properties:
        revision: { type: [string, 'null'], pattern: '^[0-9a-f]{40}$', description: "commit checked out; null if the git stage failed" }
        status: { $ref: '#/components/schemas/ApplyStatus' }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        error_summary: { type: string, maxLength: 1024, description: "first failed task + msg from the ansible JSON callback, or the git error" }
        log_tail: { type: string, maxLength: 65536, description: "stored truncated to 8 KiB head+tail; ignored on success" }
        agent_version: { type: string, maxLength: 40 }
      examples:
        - revision: "b89f176e2c1c4d1f0a9e8d7c6b5a49382716f5e4"
          status: failed
          started_at: "2026-09-30T20:00:05Z"
          finished_at: "2026-09-30T20:00:41Z"
          error_summary: "TASK [base : install packages] — apt: Unable to locate package foo"
          log_tail: "PLAY [localhost] ***…"
          agent_version: "0.1.0"
    ReportAck:
      type: object
      required: [id, received_at, transition]
      properties:
        id: { type: string, format: uuid }
        received_at: { type: string, format: date-time }
        transition: { type: [string, 'null'], enum: [failed, recovered, null] }
      examples:
        - { id: "0a1b2c3d-4e5f-4a6b-8c7d-9e0f1a2b3c4d", received_at: "2026-09-30T20:00:42Z", transition: failed }
    EscrowRequest:
      oneOf:
        - type: object
          additionalProperties: false
          required: [kind, username, ciphertext, rotated_at, escrow_key_id]
          properties:
            kind: { type: string, const: local_admin }
            username: { type: string, pattern: '^[A-Za-z0-9_.-]{1,32}$' }
            ciphertext: { type: string, pattern: '^[A-Za-z0-9+/]+={0,2}$', maxLength: 1400 }
            rotated_at: { type: string, format: date-time }
            escrow_key_id: { type: string, pattern: '^[0-9a-f]{64}$' }
        - type: object
          additionalProperties: false
          required: [kind, label, ciphertext, escrow_key_id]
          properties:
            kind: { type: string, const: luks_recovery }
            label: { type: string, pattern: '^[A-Za-z0-9_./:-]{1,64}$', description: "volume or keyslot identifier, e.g. `/` or `nvme0n1p3:slot1`" }
            ciphertext: { type: string, pattern: '^[A-Za-z0-9+/]+={0,2}$', maxLength: 1400 }
            escrow_key_id: { type: string, pattern: '^[0-9a-f]{64}$' }
      examples:
        - { kind: local_admin, username: opale-admin, ciphertext: "Zm9vYmFyYmF6cXV4Zm9vYmFyYmF6cXV4", rotated_at: "2026-09-30T20:00:00Z", escrow_key_id: "0f3c4d5e6f7a8b9c0d1e2f3a4b5c6d7e8f9a0b1c2d3e4f5a6b7c8d9e0f1a2b3c" }
        - { kind: luks_recovery, label: "/", ciphertext: "Zm9vYmFyYmF6cXV4Zm9vYmFyYmF6cXV4", escrow_key_id: "0f3c4d5e6f7a8b9c0d1e2f3a4b5c6d7e8f9a0b1c2d3e4f5a6b7c8d9e0f1a2b3c" }
    EscrowAck:
      type: object
      required: [kind, key_id, stored_at]
      properties:
        kind: { type: string, enum: [local_admin, luks_recovery] }
        id: { type: string, format: uuid, description: "recovery key row id (luks_recovery only)" }
        key_id: { type: string }
        stored_at: { type: string, format: date-time }
      examples:
        - { kind: luks_recovery, id: "0a1b2c3d-4e5f-4a6b-8c7d-9e0f1a2b3c4d", key_id: "0f3c4d5e6f7a8b9c0d1e2f3a4b5c6d7e8f9a0b1c2d3e4f5a6b7c8d9e0f1a2b3c", stored_at: "2026-09-30T20:00:43Z" }

    # ── admin objects ──
...